requirements.txt
openai==1.3.5
python-telegram-bot==20.6
requests==2.31.0
python-dotenv==1.0.0